cr + acv: #COR-2244 Deletion with plain object member works correctly.

+++ db4oj/core/src/com/db4o/internal/OpenTypeHandler.java
@@ -52,7 +52,9 @@ public class OpenTypeHandler implements ReferenceTypeHandler, ValueTypeHandler, 
         int classMetadataID = context.readInt();
         TypeHandler4 typeHandler = container().classMetadataForID(classMetadataID).typeHandler();
         if(typeHandler != null){
-            context.delete(typeHandler);
+            if(! isPlainObject(typeHandler)){
+                context.delete(typeHandler);
+            }
         }
         context.seek(linkOffset);
 	}
